# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Telegram-clone chat API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/search:
    get:
      operationId: searchUsers
      tags: [users]
      summary: Search users by phone or username
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
          description: Phone number or username to search
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
  /users/{userId}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user by ID
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: User not found
  /users/me/profile:
    patch:
      operationId: updateProfile
      tags: [users]
      summary: Update current user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileUpdate"
      responses:
        "200":
          description: Profile updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated

  # ── Chats ─────────────────────────────────────────────────────────────────
...
  /chats/{chatId}/messages:
    get:
      operationId: getChatMessages
      tags: [messages]
      summary: List messages in a chat
      parameters:
        - name: chatId
          in: path
          required: true
          schema:
            type: integer
        - name: before
          in: query
          required: false
          schema:
            type: integer
          description: Cursor for pagination (message ID)
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: List of messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"

    post:
      operationId: sendMessage
      tags: [messages]
      summary: Send a message to a chat
      parameters:
        - name: chatId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MessageInput"
      responses:
        "201":
          description: Message sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
  /chats/{chatId}/messages/{messageId}:
    delete:
      operationId: deleteMessage
      tags: [messages]
      summary: Delete a message
      parameters:
        - name: chatId
          in: path
          required: true
          schema:
            type: integer
        - name: messageId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Message deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # ── Voice rooms ───────────────────────────────────────────────────────────
  /chats/{chatId}/voice:
    get:
      operationId: getVoiceRoom
      tags: [voice]
      summary: Get active voice/video room participants for a group chat
      parameters:
        - name: chatId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Voice room state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VoiceRoom"

    post:
      operationId: joinVoiceRoom
      tags: [voice]
      summary: Join the voice/video room of a group chat
      parameters:
        - name: chatId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VoiceJoinInput"
      responses:
        "200":
          description: Joined
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VoiceRoom"

    delete:
      operationId: leaveVoiceRoom
      tags: [voice]
      summary: Leave the voice/video room
      parameters:
        - name: chatId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Left
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message

    # ── Auth schemas ──────────────────────────────────────────────────────
...
    User:
      type: object
      required: [id, phone, username, createdAt]
      properties:
        id:
          type: integer
        phone:
          type: string
        username:
          type: string
        bio:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    ProfileUpdate:
      type: object
      properties:
        username:
          type: string
          minLength: 3
        bio:
          type: ["string", "null"]
        phone:
          type: string
        password:
          type: string
          minLength: 6
        avatarUrl:
          type: ["string", "null"]

    # ── Chat schemas ──────────────────────────────────────────────────────
...
    Message:
      type: object
      required: [id, chatId, senderId, content, createdAt]
      properties:
        id:
          type: integer
        chatId:
          type: integer
        senderId:
          type: integer
        content:
          type: string
        createdAt:
          type: string
          format: date-time
        sender:
          $ref: "#/components/schemas/User"
    MessageInput:
      type: object
      required: [content]
      properties:
        content:
          type: string
          minLength: 1

    # ── Voice schemas ─────────────────────────────────────────────────────
    VoiceRoom:
      type: object
      required: [chatId, participants, active]
      properties:
        chatId:
          type: integer
        active:
          type: boolean
        participants:
          type: array
          items:
            $ref: "#/components/schemas/VoiceParticipant"
    VoiceParticipant:
      type: object
      required: [user, micEnabled, cameraEnabled, joinedAt]
      properties:
        user:
          $ref: "#/components/schemas/User"
        micEnabled:
          type: boolean
        cameraEnabled:
          type: boolean
        joinedAt:
          type: string
          format: date-time
    VoiceJoinInput:
      type: object
      required: [micEnabled, cameraEnabled]
      properties:
        micEnabled:
          type: boolean
        cameraEnabled:
          type: boolean